+++ Sources/StreamChatUI/ChatMessageList/ChatMessage/ChatMessageContentView.swift
@@ -356,7 +356,14 @@ open class _ChatMessageContentView<ExtraData: ExtraDataTypes>: _View, ThemeProvi
             bubbleThreadMetaContainer
         ].compactMap { $0 }
 
-        if options.contains(.flipped) {
+        if options.contains(.centered) {
+            mainContainer.addArrangedSubviews([bubbleThreadMetaContainer])
+            
+            constraintsToActivate += [
+                mainContainer.centerXAnchor
+                    .pin(equalTo: centerXAnchor)
+            ]
+        } else if options.contains(.flipped) {
             mainContainer.addArrangedSubviews(mainContainerSubviews.reversed())
 
             if let errorIndicator = errorIndicatorView {
@@ -384,7 +391,6 @@ open class _ChatMessageContentView<ExtraData: ExtraDataTypes>: _View, ThemeProvi
             constraintsToActivate += [
                 mainContainer.leadingAnchor
                     .pin(equalTo: leadingAnchor)
-                    .almostRequired
             ]
         }
 
@@ -422,8 +428,22 @@ open class _ChatMessageContentView<ExtraData: ExtraDataTypes>: _View, ThemeProvi
         }
 
         // Text
+        var textColor = appearance.colorPalette.text
+        var textFont = appearance.fonts.body
+        
+        if content?.isDeleted == true {
+            textColor = appearance.colorPalette.textLowEmphasis
+        } else if content?.shouldRenderAsJumbomoji == true {
+            textFont = appearance.fonts.emoji
+        } else if content?.type == .system {
+            textFont = appearance.fonts.caption1.bold
+            textColor = appearance.colorPalette.textLowEmphasis
+        }
+        
+        textView?.textColor = textColor
+        textView?.font = textFont
         textView?.text = content?.textContent
-
+        
         // Avatar
         let placeholder = appearance.images.userAvatarPlaceholder1
         if let imageURL = content?.author.imageURL {